% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_colors.R
\name{build_colors}
\alias{build_colors}
\title{Build table of colors to use}
\usage{
build_colors()
}
\value{
data.frame
}
\description{
Create a table using the HLS framework, hue values on 0-1, 0-255, and 0-360 ranges
}
