% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_palettes.R
\name{cb}
\alias{cb}
\title{colorbuildr color names}
\usage{
cb(color = "grey3")
}
\arguments{
\item{color}{the unique color names used in the package, ex: "brightred5", "grey4", "dullblue2"}
}
\description{
colorbuildr color names
}
\examples{
cb(c("violet4", "brightcyan3"))

}
