% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_palettes.R
\name{cb_pal_across}
\alias{cb_pal_across}
\title{Generates a palette within across hues}
\usage{
cb_pal_across(palette = "ROYLCBVMW", light = 3, sat = "bright",
  return = NULL)
}
\arguments{
\item{palette}{the first letter of each hue to include}

\item{light}{the lightness value to hold constant (1:7)}

\item{sat}{the saturation value to hold constant ("bright", "muted", "dull", "")}

\item{return}{defaults to returning hex codes but can also return a table or plot of the generated palette}
}
\description{
Generates a palette within across hues
}
\examples{
cb_pal_across()
cb_pal_across(sat = "bright", return = "plot")
cb_pal_across(sat = c("bright", "muted"), return = "plot")
cb_pal_across(return = "plot")

cb_pal_across(palette = "BO", sat = "bright", return = "plot")
cb_pal_across(palette = "BO", sat = "bright", return = "table")
cb_pal_across(palette = "RYCMG", light =  c(1,4), return = "plot")

}
