% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-color_table.R
\docType{data}
\name{color_table}
\alias{color_table}
\title{Table of available colors}
\format{
A data frame with 200 observations and 15 variables
\describe{
\item{H360}{hue on a 0-360 scale}
\item{L1}{lightness on a 0-1 scale}
\item{S1}{saturation on a 0-1 scale}
\item{light}{the light value used in the package, 0-7}
\item{color}{the base color name (hue), red, cyan, etc.}
\item{letter}{the first letter of the color, for building palettes}
\item{sat}{the saturation value used in the package, "bright", "muted", "dull", or blank ""}
\item{color_sat}{the color + the saturation, ex: "brightblue", "dullred"}
\item{color_name}{the final unique name: color_sat + lightness, ex: "brightblue2", "mutedorange3"}
\item{H1}{hue on a 0-1 scale}
\item{hex}{the hex code of the color}
\item{R}{the red of the RGB value}
\item{G}{the green of the RGB value}
\item{B}{the blue of the RGB value}
\item{H255}{for convenience as some HLS selection tools use a 0-255 scale}
}
}
\usage{
color_table
}
\description{
This is a list of simplified color names
}
\keyword{datasets}
