% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_palettes.R
\name{sc}
\alias{sc}
\title{simplecolors color names}
\usage{
sc(...)
}
\arguments{
\item{...}{the unique color names used in the package, ex: "brightred5", "grey4", "dullblue2"}
}
\description{
simplecolors color names
}
\examples{
sc("violet4", "brightteal3")

}
