% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_colors.R
\name{show_colors}
\alias{show_colors}
\title{Show all available colors}
\usage{
show_colors(display_type = "colorbuildr")
}
\arguments{
\item{hue_max}{}
}
\value{
ggplot
}
\description{
Show all available colors
}
\examples{

show_colors(display_type = "colorbuildr")
show_colors(display_type = "1")
show_colors(display_type = "255")
show_colors(display_type = "260")
}
