% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_palettes.R
\name{show_palette}
\alias{show_palette}
\title{Title}
\usage{
show_palette(df = color_table)
}
\arguments{
\item{df}{passing a data frame with `H360`, `color`, `letter`, `sat`, `light` and `hex`}
}
\description{
Title
}
\examples{

show_palette(color_table)

}
