% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplecolors-package.R
\docType{package}
\name{simplecolors-package}
\alias{simplecolors}
\alias{simplecolors-package}
\title{simplecolors: A simplified method for selecting colors and
    color palettes}
\description{
Using the HLS colorspace model, colors and color
    palettes can be generated with ease.  You can build a light-to-dark
    palette in the same hue by holding the hue and saturation values
    constant and varying the lightness. Similarly, you can build a nice
    color palette across hues by holding the light and saturation values
    constant and varying the hue.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rjake/simplecolors}
  \item Report bugs at \url{https://github.com//rjake/simplecolors/issues}
}

}
\author{
\strong{Maintainer}: Jake Riley \email{rjake@sas.upenn.edu}
}
\keyword{internal}
