% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplecolors.R
\docType{package}
\name{simplecolors}
\alias{simplecolors}
\title{simplecolors: A package for accessing color names using a standardized nomenclature}
\description{
The simplecolors package provides two categories of functions:
color names and color palettes
}
\section{Color names}{

The sc() function is the main way to access color names. Colors can be called by including them as comma separated string values. For example: sc("brightblue4", "mutedred2")
}

\section{Color palettes}{

There are several functions that are prefixed with "sc_". These generate palettes of colors and can return hex codes (default), a table, or a plot showing the colors selected.
}

