% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_palettes.R
\name{specify_output}
\alias{specify_output}
\title{Helper function to print output}
\usage{
specify_output(df, return = NULL)
}
\arguments{
\item{df}{a dataframe built from color_table}

\item{return}{defaults to returning hex codes but can also return a table or plot of the generated palette}
}
\description{
Helper function to print output
}
