% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_palettes.R
\name{sc}
\alias{sc}
\title{Specify color(s) by name}
\usage{
sc(...)
}
\arguments{
\item{...}{the unique color names used in the package, ex: "brightred5", "grey4", "dullblue2"}
}
\value{
a vector of hex codes
}
\description{
Specify color(s) by name
}
\examples{
sc("violet4", "brightteal3")

}
