% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_palettes.R
\name{sc}
\alias{sc}
\title{simplecolors color names}
\usage{
sc(color = "grey3")
}
\arguments{
\item{color}{the unique color names used in the package, ex: "brightred5", "grey4", "dullblue2"}
}
\description{
simplecolors color names
}
\examples{
sc(c("violet4", "brightcyan3"))

}
