% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_colors.R
\name{show_colors}
\alias{show_colors}
\title{Show all available colors}
\usage{
show_colors(labels = FALSE)
}
\arguments{
\item{labels}{logical TRUE (default) will plot the color with color names, FALSE will plot the colors only}
}
\value{
Returns a ggplot object
}
\description{
Plots all available color values.
}
\details{
Labels can be added by using the argument \code{labels = TRUE}
}
\examples{
show_colors()
}
